yasu

+++ texture/generate_texture_floorplan.h
@@ -57,7 +57,7 @@ struct TextureInput {
   int max_texture_size_per_wall_patch;
   int texture_height_per_wall;
   double position_error_for_floor;
-  double patch_size_for_synthesis;
+  int patch_size_for_synthesis;
   int num_cg_iterations;
 };
